health_monitring/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
pyserial>=3.5
joblib>=1.1.0